criaçao do crud de produtos

+++ src/main/java/com/api/projetoFinal/repositories/ProdutoRepository.java
@@ -6,8 +6,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Long> {
+public interface ProdutoRepository extends JpaRepository<Produto, Integer> {
 
     Optional<Produto> findById(Integer id);
-
 }
